added the inner satellite STMAS method.
